+++ test/function_analyzer/function_analyzer.dart
@@ -1,5 +1,1 @@
-
-
-void functionAnalyzerTest(){
-  
-}
+void functionAnalyzerTest() {}