+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to the **Dead Code Analyzer** project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## 1.0.5
+- Fix the comment function not found and added test case for comment function handling 
+
 ## 1.0.4
 - Fix the class comment not found and added test case for class handling